% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/appopt.R
\docType{package}
\name{appopt-package}
\alias{appopt}
\alias{appopt-package}
\title{appopt: An approximate optimal threshold blocking algorithm}
\description{
An approximate optimal threshold blocking algorithm
TODO: write real description
}

