% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_block_stats.R
\name{get_block_stats}
\alias{get_block_stats}
\title{Information about blocking}
\usage{
get_block_stats(data, block_size, blocking)
}
\arguments{
\item{data}{a numeric matrix or data frame describing the covariates
of the units in the experiment.}

\item{block_size}{the desired minimum block size.}

\item{blocking}{an integer vector describing the blocking.}
}
\value{
\code{get_block_stats} returns a named numeric vector of length
  two, where the first item is the maximum distance and the second item
  is the mean.
}
\description{
This function derives the maximum and average distance among all within-block
distances for a given blocking. The average is the mean within-block
distance in each block weighted by block size. The function also checks
so each block contain at least \code{block_size} units.
}
\examples{
# Generate data sets
data <- data.frame(x1 = rnorm(200), x2 = rnorm(200))

# Get blocking
blocking <- get_blocking(data, 2)

# Get information
get_block_stats(data, 2, blocking$blocks)
}

