% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_blocking.R
\name{get_blocking}
\alias{get_blocking}
\title{Construct threshold blocking}
\usage{
get_blocking(data, block_size, algorithm = "directed", treetype = "kdtree")
}
\arguments{
\item{data}{a numeric matrix or data frame describing the covariates
of the units in the experiment.}

\item{block_size}{the desired minimum block size.}

\item{algorithm}{the version of the algorithm to use. Must be one of
"directed", "undirected" or "paper".}

\item{treetype}{the method to find nearest neighbors with. Must be one
  of "brute", "kdtree" or "bdtree". "kdtree" is recommended in most cases
  but for small samples or high dimensional covariates "brute" might perform
  better.}
}
\value{
\code{get_blocking} returns a data frame where one variable is the
  units' labels (\code{1:n}) and one variable is their derived block membership.
  Details about the used algorithm is added as attributes to the data frame.
}
\description{
This function derives an approximately optimal solution to the
threshold blocking problem where the objective is to minimize
the maximum Euclidean distances between any two units
within the same block.
}
\examples{
data <- data.frame(x1 = rnorm(200),
                   x2 = rnorm(200),
                   x3 = runif(200))

get_blocking(data, 4)
}

